# app/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "NAVA-ASSIST-BACKEND",
        "version": "0.1.0"
    },
    "paths": {
        "/v1/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Create account of user and send otp on the phone number",
                "operationId": "signup_v1_auth_signup_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignUpResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/verify": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Verify otp and return token",
                "operationId": "test_token_v1_auth_verify_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyPhoneRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerifyPhoneResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login user and get user info ",
                "operationId": "login_v1_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/refresh-token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "operationId": "refresh_token_handler_v1_auth_refresh_token_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshTokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/completions": {
            "post": {
                "tags": [
                    "code"
                ],
                "summary": "Completion",
                "operationId": "completion_v1_completions_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompletionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompletionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/chat": {
            "post": {
                "tags": [
                    "code"
                ],
                "summary": "Chat",
                "operationId": "chat_v1_chat_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChatRequest": {
                "properties": {
                    "userQuery": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Userquery"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User"
                    },
                    "debug_options": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Debug Options"
                    }
                },
                "type": "object",
                "title": "ChatRequest"
            },
            "ChatResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "response": {
                        "type": "string",
                        "title": "Response"
                    },
                    "debug_data": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Debug Data"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "response"
                ],
                "title": "ChatResponse"
            },
            "CompletionRequest": {
                "properties": {
                    "language": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language"
                    },
                    "segments": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Segments"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User"
                    },
                    "debug_options": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Debug Options"
                    }
                },
                "type": "object",
                "title": "CompletionRequest"
            },
            "CompletionResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "choices": {
                        "items": {
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Choices"
                    },
                    "debug_data": {
                        "anyOf": [
                            {
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Debug Data"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "choices"
                ],
                "title": "CompletionResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LoginRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email",
                        "description": "user email"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 24,
                        "minLength": 5,
                        "title": "Password",
                        "description": "user password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "LoginRequest"
            },
            "RefreshTokenRequest": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    }
                },
                "type": "object",
                "required": [
                    "user_id"
                ],
                "title": "RefreshTokenRequest"
            },
            "SignUpRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email",
                        "description": "user email"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5,
                        "title": "Username",
                        "description": "user username"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 24,
                        "minLength": 5,
                        "title": "Password",
                        "description": "user password"
                    },
                    "phone": {
                        "type": "string",
                        "maxLength": 10,
                        "minLength": 10,
                        "title": "Phone",
                        "description": "user phone number"
                    },
                    "additional_info": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Additional Info"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "username",
                    "password",
                    "phone"
                ],
                "title": "SignUpRequest"
            },
            "SignUpResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "SignUpResponse"
            },
            "TokenResponse": {
                "properties": {
                    "jwt_token": {
                        "type": "string",
                        "title": "Jwt Token"
                    },
                    "expires_in": {
                        "type": "integer",
                        "title": "Expires In"
                    }
                },
                "type": "object",
                "required": [
                    "jwt_token",
                    "expires_in"
                ],
                "title": "TokenResponse"
            },
            "User": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "phone": {
                        "type": "string",
                        "title": "Phone"
                    },
                    "additional_info": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Additional Info"
                    },
                    "sessions": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sessions"
                    },
                    "token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Token"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "username",
                    "email",
                    "phone"
                ],
                "title": "User"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VerifyPhoneRequest": {
                "properties": {
                    "phone": {
                        "type": "string",
                        "maxLength": 10,
                        "minLength": 10,
                        "title": "Phone",
                        "description": "user phone number"
                    },
                    "otp": {
                        "type": "string",
                        "title": "Otp"
                    }
                },
                "type": "object",
                "required": [
                    "phone",
                    "otp"
                ],
                "title": "VerifyPhoneRequest"
            },
            "VerifyPhoneResponse": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "jwt_token": {
                        "type": "string",
                        "title": "Jwt Token"
                    },
                    "expires_in": {
                        "type": "integer",
                        "title": "Expires In"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "username",
                    "email",
                    "jwt_token",
                    "expires_in"
                ],
                "title": "VerifyPhoneResponse"
            },
...
        }
    }
}